SKERN-3045: Force removal of  preinstall-toolkit-builder image
To ensure the latest preinstall-toolkit-builder image is used for the
kiwi-ng build, any existing image that was cached must be removed.

The 'docker rmi -f' will force the removal. If any container is
currently using the image, the image is only untagged and will
be removed once containers with references have exited.