CASM-1896: Remove SLE update repos
Removed the update repos so LiveOS is using the same kernel
as the ncn-w001 installed OS. This resolved the CPU stack
traces being dumped during boot.

Similar stack traces are seen on pepsi as it was using the updated
kernel, but at this stage we want to use the same kernel as
is installed.