CASM-1949: Change network interface definition to use DHCP
The current configuration for eth0 is fixed for what was expected
on wasp-ncn-w001. This changes it to use DHCP for interface named
eth0. That will get networking operational in a VM environment.

NCNs will be using a different naming scheme for interfaces, such
as em0, em1, em2, etc. Default interface definitions will be provided for
those once the initial settings are determined. Full configuration will
be done via data to be provided in the Configuration Payload.