CASM-1957: Add boot payload= boot parameter
Add payload boot parameter that uses URI values. This parameter will
be processed by a dracut module to determine where the payload is
found and how it should be processed. The file://path will be used
to specify a device to mount. Mount point will always be /payload.
The file:// will be the only option initially supported.

Leaves option to support nfs:// or some other scheme in the future.