Fix timestamps in `PIT_SLUG` (#50)
`PIT_SLUG` was only defined during the prepare stage, since it wasn't
being set during the build stage a new timestamp was being generated.
This mean that the ISO name that printed at the beginning of the build
was incorrect. The ISO name that prints at the beginning is a maintainer
helper, so the developer/maintainer can see the ISO name before the ISO
is finished. If this is wrong then it is misleading.

For example, users may update `csm/assets.sh` prior to the ISO finishing
a build but if they use the name printed off in the console then the
ISO's final name will mismatch and `assets.sh` will be wrong.